chore: 사용하지 않는 코드 제거 및 띄어쓰기, 줄바꿈 수정
